requirements.txt
boto3==1.35.62
botocore==1.35.62
jmespath==1.0.1
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.3
setuptools==75.5.0
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
wheel==0.45.0
